requirements.txt
opencv-python
numpy
paddlepaddle
paddleocr
flask
flask-cors
Pillow